feat(angle): Add vaults on Arbitrum
