+++ src/apps/angle/angle.module.ts
@@ -2,7 +2,8 @@ import { Module } from '@nestjs/common';
 
 import { AbstractApp } from '~app/app.dynamic-module';
 
-import { AngleApiHelper } from './common/angle.api';
+import { ArbitrumAngleVaultsContractPositionFetcher } from './arbitrum/angle.vault.contract-position-fetcher';
+import { AnglePositionResolver } from './common/angle.position-resolver';
 import { AngleViemContractFactory } from './contracts';
 import { EthereumAngleSanTokenTokenFetcher } from './ethereum/angle.san-token.token-fetcher';
 import { EthereumAngleVaultsContractPositionFetcher } from './ethereum/angle.vault.contract-position-fetcher';
@@ -11,7 +12,9 @@ import { EthereumAngleVeAngleContractPositionFetcher } from './ethereum/angle.vo
 @Module({
   providers: [
     AngleViemContractFactory,
-    AngleApiHelper,
+    AnglePositionResolver,
+    // Arbitrum
+    ArbitrumAngleVaultsContractPositionFetcher,
     // Ethereum
     EthereumAngleSanTokenTokenFetcher,
     EthereumAngleVeAngleContractPositionFetcher,
